Fix linting of default and examples siblings to $ref in Draft 7 and older
In Draft 7 and older JSON Schema versions, all keywords that are siblings
to $ref are ignored by the specification. This means that default and
examples keywords next to $ref should not be validated/linted.

In Draft 2019-09 and 2020-12, $ref siblings are valid and should be linted.

This commit updates ValidDefault and ValidExamples linter rules to:
- Skip linting when $ref is present in Draft 7, Draft 6, and Draft 4
- Continue linting when $ref is present in Draft 2019-09 and 2020-12

Fixes: https://github.com/sourcemeta/jsonschema/issues/425
Co-Authored-By: unknown <>